Add most up-to-date version of the VQE simulator.
This is still in development. Ultimately, this needs be a Python module/package.
But the current Jupyter-notebook form is convenient for playing around
with the code and learning Qiskit's functionality. The main notebook is
"VQE.ipynb". Different parts of the VQE workflow have smaller notebooks.
They're not required to run the main one, but are convenient if you want
to look at an isolated element of the simulation.