Add Mathematica notebooks for exact diagonalization
savage.nb is the main notebook that performs projection onto
zero-momentum sector, truncation, and splits the resulting Hilbert space
into even and odd parity sectors. It diagonalizes the corresponding Hamiltonians,
expresses the eigenstates in the computational basis of the QC
simulation. Interpolating operators are constructed and exact effective
mass curves are computed. variational_layer.nb is a notebook to play
with the form of the variational layer in the QC simulation as an
explicit unitary. operators.nb is a notebook where the numerical matrix
form of the interpolators is computed in the (concatenated) QC
computational basis of even and odd parity sectors.